worker_node/requirements.txt
fastapi==0.78.0
uvicorn==0.17.0
cassandra-driver==3.25.0
numpy==1.23.5
scikit-learn==1.1.3
requests==2.28.1
matplotlib==3.9.2
tensorflow-cpu==2.13.0
redis==5.1.1
confluent-kafka==2.6.0